requirements.txt
crewai
crewai_tools
python-dotenv
langchain-community
langchain-openai
